refactor(SettingsSection): move to shared/ subdirectory
Although it's not actually shared yet lol
